+++ src/components/dataTable/DataTableNG.tsx
@@ -6,6 +6,8 @@ import { IconButton } from '@mui/material';
 import DeleteIcon from '@mui/icons-material/DeleteOutlined';
 import EditIcon from '@mui/icons-material/Edit';
 import { useNavigate } from 'react-router-dom';
+import Swal from 'sweetalert2';
+import { useState, useEffect  } from 'react';
 
 type Props = {
     columns: GridColDef[],
@@ -14,16 +16,36 @@ type Props = {
 }
 
 const DataTableNG = (props: Props) => {
+    const [rows, setRows] = useState(props.rows);
     const navigate = useNavigate();
 
-    const handleDelete = (IDNG: number) => {
-        axios.delete(`${host_server}/${props.slug}/${IDNG}`)
-        
-    }
+    const handleDelete = async (IDNG: number) => {
+        const result = await Swal.fire({
+            title: '¿Estás seguro de borrar?',
+            text: "¡No podrás revertir esto!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Sí, bórralo!'
+        });
+        if (result.isConfirmed) {
+            await axios.delete(`${host_server}/${props.slug}/${IDNG}`);
+            Swal.fire('Borrado!', 'El registro ha sido borrado.', 'success');
+            setRows(rows.filter((row: any) => row.IDNG !== IDNG));
+        }
+    };
+
+
     const handleEditClick = (IDNG: number) => {
         navigate(`/EditarPersonasReportadas/${IDNG}`);
     }
 
+    useEffect(() => {
+        setRows(props.rows);
+    }, [props.rows]);
+
+    
     const actionColumn: GridColDef = {
         field: 'acciones',
         headerName: 'Acciones',
@@ -65,6 +87,9 @@ const DataTableNG = (props: Props) => {
                             pageSize: 10,
                         },
                     },
+                }}
+                localeText={{
+                    noRowsLabel: 'No hay registros',
                 }}
                 slots={{ toolbar: GridToolbar }}
                 slotProps={{
